backend/requirements.txt
fastapi
uvicorn
pdfplumber
fpdf2
python-multipart
